Diffusion cleaned up
